+++ tests/unit/connector/test_elasticsearch_output.py
@@ -216,7 +216,7 @@ class TestElasticsearchOutput(BaseOutputTestCase):
                 {"anything": "anything"},
                 [{"foo": "bar"}, {"bar": "baz"}],
                 0,
-                search.exceptions.TransportError,
+                FatalOutputError,
             ),
             (
                 429,
@@ -240,7 +240,7 @@ class TestElasticsearchOutput(BaseOutputTestCase):
                 {"anything": "anything"},
                 [{"foo": "*" * 500}],
                 1,
-                search.exceptions.TransportError,
+                FatalOutputError,
             ),
             (
                 429,
@@ -248,7 +248,7 @@ class TestElasticsearchOutput(BaseOutputTestCase):
                 {"anything": "anything"},
                 [{"foo": "*" * 500}],
                 1,
-                search.exceptions.TransportError,
+                FatalOutputError,
             ),
             (
                 429,
@@ -264,7 +264,7 @@ class TestElasticsearchOutput(BaseOutputTestCase):
                 {"error": {"reason": "wrong_reason"}},
                 [{"foo": "*" * 500}],
                 1,
-                search.exceptions.TransportError,
+                FatalOutputError,
             ),
             (
                 429,
@@ -292,7 +292,7 @@ class TestElasticsearchOutput(BaseOutputTestCase):
                 },
                 [{"foo": "*" * 500}],
                 1,
-                search.exceptions.TransportError,
+                FatalOutputError,
             ),
         ],
     )